Docs, lints + test fixes
