+++ tamr_toolbox/notifications/slack.py
@@ -1,21 +1,23 @@
 import logging
-import os
 import warnings
 from abc import ABC
 from dataclasses import dataclass
 from typing import Dict
 from typing import Union, List, Optional
 
-import slack
-from slack.errors import SlackApiError
 from tamr_unify_client import Client
 from tamr_unify_client.operation import Operation
 
 from tamr_toolbox.models.data_type import JsonDict
 from tamr_toolbox.models.operation_state import OperationState
-from tamr_toolbox.notifications.common import _monitor_job as monitor_job_common
-from tamr_toolbox.notifications.core import _BaseNotifier
+from tamr_toolbox.notifications.common import _monitor_job as monitor_job_common, _BaseNotifier
 from tamr_toolbox.utils.operation import get_details
+
+try:
+    import slack
+    from slack.errors import SlackApiError
+except ImportError:
+    pass
 
 LOGGER = logging.getLogger(__name__)
 
@@ -53,7 +55,7 @@ class SlackNotifier(_BaseNotifier, ABC):
                 channel=self.channel,
                 text=message,
                 username="Tamr Notifications Bot",
-                icon_url="https://jdp491bprdv1ar3uk2puw37i-wpengine.netdna-ssl.com/wp-content/uploads/2020/08/Tamr-Square-Dark.png", # noqa
+                icon_url="https://jdp491bprdv1ar3uk2puw37i-wpengine.netdna-ssl.com/wp-content/uploads/2020/08/Tamr-Square-Dark.png",  # noqa
             )
         except SlackApiError as e:
             LOGGER.error(f"Error posting message: {e}.")
@@ -62,12 +64,12 @@ class SlackNotifier(_BaseNotifier, ABC):
 
 
 def send_message(
-        *,
-        slack_client: "slack.WebClient",
-        channel: str,
-        message: str,
-        raise_error: bool = True,
-        **kwargs,
+    *,
+    slack_client: "slack.WebClient",
+    channel: str,
+    message: str,
+    raise_error: bool = True,
+    **kwargs,
 ) -> JsonDict:
     """Sends a message to a pre-defined Slack channel
 
@@ -87,8 +89,12 @@ def send_message(
 
     # This function requires slack, an optional dependency
     from slack.errors import SlackApiError
-    warnings.warn("'This function will be deprecated in a future release. "
-                  "Use SlackNotifier().send_message() instead.'", DeprecationWarning)
+
+    warnings.warn(
+        "'This function will be deprecated in a future release. "
+        "Use SlackNotifier().send_message() instead.'",
+        DeprecationWarning,
+    )
     try:
         slack_response = slack_client.chat_postMessage(channel=channel, text=message)
     except SlackApiError as e:
@@ -108,11 +114,11 @@ def send_message(
 
 
 def _send_job_status_message(
-        *,
-        slack_client: "slack.WebClient",
-        channel: str,
-        operation: Operation,
-        notify_states: List[OperationState],
+    *,
+    slack_client: "slack.WebClient",
+    channel: str,
+    operation: Operation,
+    notify_states: List[OperationState],
 ) -> JsonDict:
     """Checks operation state and if in `notify_states` sends the message.
 
@@ -125,8 +131,11 @@ def _send_job_status_message(
     Returns:
         The data property of the Slack WebClient object
     """
-    warnings.warn("'This function will be deprecated in a future release. "
-                  "Use slack.SlackNotifier().monitor_job() instead.'", DeprecationWarning)
+    warnings.warn(
+        "'This function will be deprecated in a future release. "
+        "Use slack.SlackNotifier().monitor_job() instead.'",
+        DeprecationWarning,
+    )
     state = OperationState[operation.state]
     if state in notify_states:
         message = get_details(operation=operation)
@@ -134,14 +143,14 @@ def _send_job_status_message(
 
 
 def monitor_job(
-        tamr: Client,
-        *,
-        slack_client: "slack.WebClient",
-        channel: str,
-        operation: Union[int, str, Operation],
-        poll_interval_seconds: float = 1,
-        timeout_seconds: Optional[float] = None,
-        notify_states: Optional[List[OperationState]] = None,
+    tamr: Client,
+    *,
+    slack_client: "slack.WebClient",
+    channel: str,
+    operation: Union[int, str, Operation],
+    poll_interval_seconds: float = 1,
+    timeout_seconds: Optional[float] = None,
+    notify_states: Optional[List[OperationState]] = None,
 ) -> List[JsonDict]:
     """Monitors a Tamr Operation and sends a Slack message to a channel when the job status is updated
 
@@ -158,8 +167,11 @@ def monitor_job(
         A list of messages with their config sent via the Slack WebClient object
     """
 
-    warnings.warn("'This function will be deprecated in a future release. "
-                  "Use SlackNotifier().monitor_job() instead.'", DeprecationWarning)
+    warnings.warn(
+        "'This function will be deprecated in a future release. "
+        "Use SlackNotifier().monitor_job() instead.'",
+        DeprecationWarning,
+    )
     list_responses = monitor_job_common(
         tamr=tamr,
         send_message=send_message,
