+++ tests/notifications/test_emails.py
@@ -14,13 +14,14 @@ CONFIG = utils.config.from_yaml(
 
 @patch("smtplib.SMTP", autospec=True)
 def test_tls_login(mock_smtp):
-    notifier = EmailNotifier(
+    _ = EmailNotifier(
         sender_address=CONFIG["my_email_notification"]["sender_address"],
         sender_password=CONFIG["my_email_notification"]["sender_password"],
         recipient_addresses=CONFIG["my_email_notification"]["recipient_addresses"],
         smtp_server=CONFIG["my_email_notification"]["smtp_server"],
         smtp_port=CONFIG["my_email_notification"]["smtp_port"],
-        use_tls=True)
+        use_tls=True,
+    )
 
     context = mock_smtp.return_value
 
@@ -31,13 +32,14 @@ def test_tls_login(mock_smtp):
 
 @patch("smtplib.SMTP_SSL", autospec=True)
 def test_non_tls_login(mock_smtp):
-    notifier = EmailNotifier(
+    _ = EmailNotifier(
         sender_address=CONFIG["my_email_notification"]["sender_address"],
         sender_password=CONFIG["my_email_notification"]["sender_password"],
         recipient_addresses=CONFIG["my_email_notification"]["recipient_addresses"],
         smtp_server=CONFIG["my_email_notification"]["smtp_server"],
         smtp_port=CONFIG["my_email_notification"]["smtp_port"],
-        use_tls=False)
+        use_tls=False,
+    )
 
     context = mock_smtp.return_value
 
@@ -51,11 +53,11 @@ def test_build_message(mock_smtp):
     test_message = "This is a test email."
     test_subject_line = "Test 123"
     expected_msg = (
-            'Content-Type: text/plain; charset="us-ascii"\nMIME-Version: '
-            + "1.0\nContent-Transfer-Encoding: 7bit\n"
-            + f'Subject: Tamr System: Test 123\nFrom: {CONFIG["my_email_notification"]["sender_address"]}\n'
-            + f'To: {CONFIG["my_email_notification"]["recipient_addresses"][0]}\n'
-            + "\nThis is a test email."
+        'Content-Type: text/plain; charset="us-ascii"\nMIME-Version: '
+        + "1.0\nContent-Transfer-Encoding: 7bit\n"
+        + f'Subject: Tamr System: Test 123\nFrom: {CONFIG["my_email_notification"]["sender_address"]}\n'  # noqa
+        + f'To: {CONFIG["my_email_notification"]["recipient_addresses"][0]}\n'
+        + "\nThis is a test email."
     )
 
     notifier = EmailNotifier(
@@ -64,7 +66,8 @@ def test_build_message(mock_smtp):
         recipient_addresses=CONFIG["my_email_notification"]["recipient_addresses"],
         smtp_server=CONFIG["my_email_notification"]["smtp_server"],
         smtp_port=CONFIG["my_email_notification"]["smtp_port"],
-        use_tls=True)
+        use_tls=True,
+    )
 
     msg = notifier._build_message(test_message, test_subject_line)
 
@@ -82,7 +85,8 @@ def test_send_message(mock_smtp):
         recipient_addresses=CONFIG["my_email_notification"]["recipient_addresses"],
         smtp_server=CONFIG["my_email_notification"]["smtp_server"],
         smtp_port=CONFIG["my_email_notification"]["smtp_port"],
-        use_tls=True)
+        use_tls=True,
+    )
 
     context = mock_smtp.return_value
     msg = notifier._build_message(test_message, test_subject_line)
@@ -91,7 +95,8 @@ def test_send_message(mock_smtp):
     context.sendmail.assert_called_with(
         CONFIG["my_email_notification"]["sender_address"],
         CONFIG["my_email_notification"]["recipient_addresses"],
-        msg)
+        msg,
+    )
 
     assert notifier.sent_messages == [msg]
 
@@ -100,11 +105,11 @@ def test_deprecated_build_message():
     test_message = "This is a test email."
     subject_line = "Test 123"
     test_response = (
-            'Content-Type: text/plain; charset="us-ascii"\nMIME-Version: '
-            + "1.0\nContent-Transfer-Encoding: 7bit\n"
-            + f'Subject: Test 123\nFrom: {CONFIG["my_email_notification"]["sender_address"]}\n'
-            + f'To: {CONFIG["my_email_notification"]["recipient_addresses"][0]}\n'
-            + "\nThis is a test email."
+        'Content-Type: text/plain; charset="us-ascii"\nMIME-Version: '
+        + "1.0\nContent-Transfer-Encoding: 7bit\n"
+        + f'Subject: Test 123\nFrom: {CONFIG["my_email_notification"]["sender_address"]}\n'
+        + f'To: {CONFIG["my_email_notification"]["recipient_addresses"][0]}\n'
+        + "\nThis is a test email."
     )
 
     msg = tbox.notifications.emails._build_message(
